refactor(openmetrics): extract shared types into types/ and dedupe IPC contracts
